+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180307134226) do
+ActiveRecord::Schema.define(version: 20180314031010) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -161,18 +161,25 @@ ActiveRecord::Schema.define(version: 20180307134226) do
     t.index ["project_id", "profile_id"], name: "index_profiles_projects_on_project_id_and_profile_id", unique: true
   end
 
+  create_table "project_setups", force: :cascade do |t|
+    t.bigint "project_id", null: false
+    t.boolean "is_completed", default: false, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["project_id"], name: "index_project_setups_on_project_id", unique: true
+  end
+
   create_table "projects", force: :cascade do |t|
     t.string "title", null: false
-    t.string "owner_type", null: false
-    t.bigint "owner_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.citext "slug", null: false
     t.text "description"
     t.citext "tags", default: [], array: true
     t.boolean "is_public", default: false, null: false
-    t.index ["owner_type", "owner_id", "slug"], name: "index_projects_on_owner_type_and_owner_id_and_slug", unique: true
-    t.index ["owner_type", "owner_id"], name: "index_projects_on_owner_type_and_owner_id"
+    t.bigint "owner_id", null: false
+    t.index ["owner_id", "slug"], name: "index_projects_on_owner_id_and_slug", unique: true
+    t.index ["owner_id"], name: "index_projects_on_owner_id"
   end
 
   create_table "revisions", force: :cascade do |t|
@@ -219,6 +226,7 @@ ActiveRecord::Schema.define(version: 20180307134226) do
   add_foreign_key "file_resources", "file_resource_snapshots", column: "current_snapshot_id"
   add_foreign_key "file_resources", "file_resources", column: "parent_id"
   add_foreign_key "profiles", "accounts"
+  add_foreign_key "project_setups", "projects"
   add_foreign_key "revisions", "profiles", column: "author_id"
   add_foreign_key "revisions", "projects"
   add_foreign_key "revisions", "revisions", column: "parent_id"