+++ spec/controllers/shared_examples/raise_404_if_non_existent.rb
@@ -2,8 +2,10 @@
 
 # Expect ActiveRecord::RecordNotFound error (404) when object does not exist
 RSpec.shared_examples 'raise 404 if non-existent' do |object|
-  context "when #{object.to_s.downcase} does not exist" do
-    before { object&.destroy_all }
-    it { expect { run_request }.to raise_error ActiveRecord::RecordNotFound }
+  let(:object_class) { object }
+  before { object_class&.destroy_all }
+
+  it "raises 404 when #{object.to_s.downcase} does not exist" do
+    expect { run_request }.to raise_error ActiveRecord::RecordNotFound
   end
 end