+++ spec/models/project_spec.rb
@@ -8,12 +8,13 @@ RSpec.describe Project, type: :model do
   end
 
   describe 'associations' do
-    it { is_expected.to belong_to(:owner) }
+    it { is_expected.to belong_to(:owner).class_name('Profiles::Base') }
     it do
       is_expected
         .to have_and_belong_to_many(:collaborators)
         .class_name('Profiles::User').validate(false)
     end
+    it { is_expected.to have_one(:setup).dependent(:destroy) }
     it do
       is_expected
         .to have_one(:staged_root_folder)
@@ -77,8 +78,21 @@ RSpec.describe Project, type: :model do
 
   describe 'attributes' do
     it { is_expected.to have_readonly_attribute(:owner_id) }
-    it { is_expected.to have_readonly_attribute(:owner_type) }
-    it { is_expected.to have_readonly_attribute(:owner_type) }
+  end
+
+  describe 'delegations' do
+    it do
+      is_expected
+        .to delegate_method(:in_progress?)
+        .to(:setup)
+        .with_prefix
+    end
+    it do
+      is_expected
+        .to delegate_method(:completed?)
+        .to(:setup)
+        .with_prefix
+    end
   end
 
   describe 'callbacks' do
@@ -98,90 +112,21 @@ RSpec.describe Project, type: :model do
         it { is_expected.not_to receive(:generate_slug_from_title) }
       end
     end
-
-    context 'after save' do
-      subject(:project)     { build(:project) }
-      let(:link_to_folder)  { 'https://drive.google.com/drive/folders/test' }
-
-      before do
-        allow(subject).to receive(:import_google_drive_folder)
-        allow(subject).to receive(:link_to_google_drive_is_accessible_folder)
-      end
-
-      context 'when import_google_drive_folder_on_save is true' do
-        before do
-          project.import_google_drive_folder_on_save = true
-          project.link_to_google_drive_folder = link_to_folder
-        end
-        after { project.save }
-        it    { is_expected.to receive(:import_google_drive_folder) }
-      end
-
-      context 'when import_google_drive_folder_on_save was true' do
-        before do
-          project.import_google_drive_folder_on_save = true
-          project.link_to_google_drive_folder = link_to_folder
-          project.save
-        end
-        it { expect(project.import_google_drive_folder_on_save).to be false }
-      end
-    end
   end
 
   describe 'validations' do
     it do
       is_expected.to validate_presence_of(:owner).with_message 'must exist'
     end
-    it do
-      is_expected
-        .to validate_inclusion_of(:owner_type).in_array %w[Profiles::Base]
-    end
     it { is_expected.to validate_presence_of(:title) }
     it { is_expected.to validate_length_of(:title).is_at_most(50) }
 
     context 'when validating slug' do
-      # Test does not work with polymorphic associations, use custom test below
-      # instead
-      # it do
-      #   is_expected
-      #     .to validate_uniqueness_of(:slug)
-      #     .case_insensitive
-      #     .scoped_to(:owner_type, :owner_id)
-      # end
-      context 'uniqueness of slug scoped to owner type + ID' do
-        let!(:first_project)      { create :project, slug: slug, owner: owner }
-        let(:slug)  { 'my-slug' }
-        let(:owner) { create(:user) }
-
-        context 'when owner type is identical but ID is different' do
-          subject(:second_project)  { build :project, slug: slug }
-          before                    { second_project.valid? }
-          it 'does not add a :slug error' do
-            expect(second_project.errors[:slug])
-              .not_to include 'has already been taken'
-          end
-        end
-
-        context 'when owner ID is identical but type is different' do
-          subject(:second_project) { build :project, slug: slug, owner: owner }
-          before do
-            second_project.owner_type = 'Project'
-            second_project.valid?
-          end
-          it 'does not add a :slug error' do
-            expect(second_project.errors[:slug])
-              .not_to include 'has already been taken'
-          end
-        end
-
-        context 'when owner ID + type are identical' do
-          subject(:second_project)  { build :project, slug: slug, owner: owner }
-          before                    { second_project.valid? }
-          it 'adds a :slug error' do
-            expect(second_project.errors[:slug])
-              .to include 'has already been taken'
-          end
-        end
+      it do
+        is_expected
+          .to validate_uniqueness_of(:slug)
+          .scoped_to(:owner_id)
+          .case_insensitive
       end
 
       it do
@@ -212,33 +157,6 @@ RSpec.describe Project, type: :model do
     end
   end
 
-  describe '.find' do
-    let!(:project) { create(:project) }
-    subject(:method) { Project.find(project.owner.to_param, project.slug) }
-
-    it 'finds project by profile handle and slug' do
-      is_expected.to eq project
-    end
-
-    context 'when profile does not exist' do
-      before { project.owner.destroy }
-      it { expect { method }.to raise_error ActiveRecord::RecordNotFound }
-    end
-
-    context 'when project does not exist' do
-      before { project.destroy }
-      it { expect { method }.to raise_error ActiveRecord::RecordNotFound }
-    end
-
-    context 'when project slug is not passed' do
-      subject(:method) { Project.find(project.id) }
-
-      it 'finds project by ID' do
-        is_expected.to eq project
-      end
-    end
-  end
-
   describe 'revisions#create_draft_and_commit_files!' do
     subject(:method) do
       project.revisions.create_draft_and_commit_files!('author')
@@ -252,60 +170,24 @@ RSpec.describe Project, type: :model do
     end
   end
 
-  describe '#import_google_drive_folder', isolated_unit_test: true do
-    subject(:method)    { project.send :import_google_drive_folder }
-    let(:project)       { build_stubbed(:project) }
-    let(:file)          { instance_double Google::Apis::DriveV3::File }
-    let(:mime_type)     { Providers::GoogleDrive::MimeType.folder }
-    let(:root_folder)   { instance_double FileResource }
+  describe '#setup_not_started?' do
+    subject(:setup_started) { project.setup_not_started? }
+    let(:not_started)       { false }
+    let(:setup)             { instance_double Project::Setup }
 
     before do
-      allow(project).to receive(:google_drive_folder_id).and_return 'folder-id'
-      allow(project).to receive(:root_folder=)
-      allow(FileResources::GoogleDrive)
-        .to receive(:find_or_initialize_by)
-        .with(external_id: 'folder-id')
-        .and_return(root_folder)
-      allow(root_folder).to receive(:pull)
-      allow(project).to receive(:root_folder).and_return root_folder
-      allow(root_folder).to receive(:id).and_return 'the-id'
-      allow(FolderImportJob).to receive(:perform_later)
+      allow(project).to receive(:setup).and_return setup
+      allow(setup).to receive(:not_started?).and_return(not_started) if setup
     end
 
-    it 'calls #pull on root folder' do
-      expect(root_folder).to receive(:pull)
-      subject
+    context 'when setup is nil' do
+      let(:setup) { nil }
+      it          { is_expected.to eq true }
     end
 
-    it 'sets root folder' do
-      expect(project).to receive(:root_folder=).with(root_folder)
-      subject
-    end
-
-    it 'creates a FolderImportJob' do
-      expect(FolderImportJob).to receive(:perform_later)
-        .with(reference: project, file_resource_id: 'the-id')
-      subject
-    end
-
-    context 'when error is raised' do
-      let(:staged_root_folder) { instance_double StagedFile }
-
-      before do
-        allow(project).to receive(:root_folder).and_raise StandardError
-        allow(project)
-          .to receive(:staged_root_folder).and_return staged_root_folder
-        allow(staged_root_folder).to receive(:destroy)
-      end
-
-      it 'calls #destroy on staged root folder' do
-        expect(staged_root_folder).to receive(:destroy)
-        expect { subject }.to raise_error StandardError
-      end
-
-      it 're-raises the error' do
-        expect { subject }.to raise_error StandardError
-      end
+    context 'when setup is present' do
+      let(:not_started) { 'value' }
+      it                { is_expected.to eq 'value' }
     end
   end